Organize data and receipt logic into their own files
