+++ menu.js
@@ -1,113 +1,8 @@
-import { saveToStorage, getCartFromStorage, getFieldFromStorage } from "./storage.js";
-import { menuItems, cart } from "./data.js";
-import * as data from './data.js';
-data.setCart(getCartFromStorage());
-data.setSubtotal(getFieldFromStorage("subtotal"));
-data.setDiscounts(getFieldFromStorage("discounts"));
-data.setTip(getFieldFromStorage("tip"));
-data.setTotal(getFieldFromStorage("total"));
-for (const foodItem in data.cart) {
-    createRow(foodItem);
-}
+import { loadData, saveData, setTip } from './data.js';
+import { renderTotal, add, clear, renderReceipt } from "./receipt.js";
+loadData();
+renderReceipt();
 renderTotal();
-function add(itemId) {
-    if (!cart[itemId]) {
-        cart[itemId] = { ...menuItems[itemId], quantity: 0 }; // Prevent unintended mutations
-    }
-    cart[itemId].quantity++;
-    data.addToSubtotal(cart[itemId].price);
-    calculateTotal();
-    updateOrCreateRow(itemId);
-}
-function updateQuantity(itemId, newQuantity) {
-    const oldQuantity = cart[itemId].quantity;
-    const price = cart[itemId].price;
-    if (newQuantity === 0) {
-        remove(itemId);
-    }
-    else {
-        cart[itemId].quantity = newQuantity;
-        updateRow(itemId);
-    }
-    data.addToSubtotal((newQuantity - oldQuantity) * price);
-    calculateTotal();
-}
-function remove(itemId) {
-    delete cart[itemId];
-    const row = getRowById(itemId);
-    if (row)
-        row.remove();
-}
-function clear() {
-    data.emptyCart();
-    const receiptTableBody = document.getElementById("receipt-table-body");
-    receiptTableBody.replaceChildren(); // Clear all rows
-    data.setSubtotal(0);
-    calculateTotal();
-}
-function updateOrCreateRow(itemId) {
-    const row = getRowById(itemId);
-    if (row) {
-        updateRow(itemId);
-    }
-    else {
-        createRow(itemId);
-    }
-}
-function updateRow(itemId) {
-    const foodItem = cart[itemId];
-    const row = getRowById(itemId);
-    if (row) {
-        row.querySelector(".item-quantity").textContent = `Quantity: ${foodItem.quantity}`;
-        row.querySelector(".item-price").textContent = `$${(foodItem.price * foodItem.quantity).toFixed(2)}`;
-    }
-}
-function createRow(itemId) {
-    const foodItem = cart[itemId];
-    const receiptTableBody = document.getElementById("receipt-table-body");
-    const newRow = document.createElement("tr");
-    newRow.setAttribute("data-id", itemId);
-    newRow.innerHTML = `
-        <td class="receipt-cell">
-            <div class="item-container">
-                <img src="${foodItem.image}" alt="${foodItem.name}" class="food-image">
-                <div class="item-quantity">Quantity: ${foodItem.quantity}</div>
-                <button class="edit-button">Edit</button>
-                <div class="item-name">${foodItem.name}</div>
-                <div class="item-price">$${(foodItem.price * foodItem.quantity).toFixed(2)}</div>
-            </div>
-        </td>
-    `;
-    // Add event listener for the edit button
-    const editButton = newRow.querySelector(".edit-button");
-    editButton.addEventListener("click", () => {
-        const newQuantity = parseInt(prompt("Enter new quantity (0 to remove):") || "0", 10);
-        if (!isNaN(newQuantity)) {
-            updateQuantity(itemId, newQuantity);
-        }
-    });
-    receiptTableBody.appendChild(newRow);
-}
-function getRowById(itemId) {
-    return document.querySelector(`tr[data-id="${itemId}"]`);
-}
-function calculateTotal() {
-    data.updateTotal();
-    renderTotal();
-}
-function renderTotal() {
-    document.getElementById("subtotal-field").innerText = `$${data.getSubtotal().toFixed(2)}`;
-    document.getElementById("discounts-field").innerText = `$${data.getDiscounts().toFixed(2)}`;
-    document.getElementById("tip-field").innerText = `$${data.getTip().toFixed(2)}`;
-    document.getElementById("total-field").innerText = `$${data.getTotal().toFixed(2)}`;
-}
-function saveData() {
-    saveToStorage("cart", data.cart);
-    saveToStorage("subtotal", data.getSubtotal());
-    saveToStorage("discounts", data.getDiscounts());
-    saveToStorage("tip", data.getTip());
-    saveToStorage("total", data.getTotal());
-}
 // Show the popup and disable main content
 document.getElementById("follower-card")?.addEventListener("click", () => {
     document.getElementById("follower-popup").classList.remove("hidden");
@@ -126,9 +21,7 @@ document.querySelector("#tip-popup button")?.addEventListener("click", () => {
     document.getElementById("tip-popup").classList.add("hidden");
     document.querySelector(".main-screen").classList.remove("disabled");
     const newTip = parseInt(document.querySelector(".slider").value);
-    data.setTip(newTip);
-    calculateTotal();
-    renderTotal();
+    setTip(newTip);
 });
 const slider = document.querySelector(".slider");
 slider.addEventListener("input", (event) => {
